[김형준] 서버 class 로 완전히 변경
protocol.h 의 몇가지 함수와 변수가 충돌을 일으켜, 대부분 대거 수정하였다. 차후 문제발생할 소지가 있으므로 꼭
참고하기바람
