+++ server/Server/Server/Debug/Server.log
@@ -1,8 +1,14 @@
-﻿  ServerMain.cpp
-d:\졸업작품\server\server\server\servermain.cpp(78): warning C4996: 'gethostbyname': Use getaddrinfo() or GetAddrInfoW() instead or define _WINSOCK_DEPRECATED_NO_WARNINGS to disable deprecated API warnings
+﻿  Main.cpp
+  코드를 생성하고 있습니다...
+  컴파일하고 있습니다...
+  iocpServerClass.cpp
+d:\program files (x86)\vs projects\졸업작품\server\server\server\iocpserverclass.cpp(33): warning C4996: 'gethostbyname': Use getaddrinfo() or GetAddrInfoW() instead or define _WINSOCK_DEPRECATED_NO_WARNINGS to disable deprecated API warnings
   c:\program files (x86)\windows kits\8.1\include\um\winsock2.h(2238): note: 'gethostbyname' 선언을 참조하십시오.
-d:\졸업작품\server\server\server\servermain.cpp(79): warning C4996: 'strcpy': This function or variable may be unsafe. Consider using strcpy_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
+d:\program files (x86)\vs projects\졸업작품\server\server\server\iocpserverclass.cpp(34): warning C4996: 'strcpy': This function or variable may be unsafe. Consider using strcpy_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
   c:\program files (x86)\windows kits\10\include\10.0.10240.0\ucrt\string.h(119): note: 'strcpy' 선언을 참조하십시오.
-d:\졸업작품\server\server\server\servermain.cpp(278): warning C4996: 'inet_ntoa': Use inet_ntop() or InetNtop() instead or define _WINSOCK_DEPRECATED_NO_WARNINGS to disable deprecated API warnings
+d:\program files (x86)\vs projects\졸업작품\server\server\server\iocpserverclass.cpp(208): warning C4996: 'inet_ntoa': Use inet_ntop() or InetNtop() instead or define _WINSOCK_DEPRECATED_NO_WARNINGS to disable deprecated API warnings
   c:\program files (x86)\windows kits\8.1\include\um\winsock2.h(1868): note: 'inet_ntoa' 선언을 참조하십시오.
-  Server.vcxproj -> D:\졸업작품\server\Server\Debug\Server.exe
+  ProcessPacket_functioncpp.cpp
+  코드를 생성하고 있습니다...
+  Server.vcxproj -> D:\Program Files (x86)\VS projects\졸업작품\server\Server\Debug\Server.exe
+  Server.vcxproj -> D:\Program Files (x86)\VS projects\졸업작품\server\Server\Debug\Server.pdb (Full PDB)
