+++ server/Server/Server/Debug/Server.tlog/Server.lastbuildstate
@@ -1,2 +1,2 @@
 #TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
-Debug|Win32|D:\졸업작품\server\Server\|
+Debug|Win32|D:\Program Files (x86)\VS projects\졸업작품\server\Server\|
